// src/common/enums/contact.enums.ts
export enum ContactRole {
  COLLABORATOR = "collaborator",
  PROVIDER = "provider",
  TENANT = "tenant",
}

export enum ContactType {
  INDIVIDUAL = "INDIVIDUAL",
  COMPANY = "COMPANY",
}

export enum ContactDetailKey {
  BIRTH_DATE = "birth_date",
  LEGAL_NAME = "legal_name",
  TRADE_NAME = "trade_name",
  STATE_REGISTRATION = "state_registration",
  MUNICIPAL_REGISTRATION = "municipal_registration",
  ZIP_CODE = "zip_code",
  STATE = "state",
  CITY = "city",
  STREET = "street",
  COMPLEMENT = "complement",
  TAX_ID = "tax_id",
  NOTES = "notes",
}