// src/common/enums/subscription.enum.ts
export enum SubscriptionStatus {
  PENDING = "PENDING",
  ACTIVE = "ACTIVE",
  PAST_DUE = "PAST_DUE",
  INACTIVE = "INACTIVE",
}

export enum AttachmentOwnerType {
  USER = "user",
  CONTACT = "contact",
  PROJECT = "project",
  AMENITY = "amenity",
  TASK = "task",
  TRANSACTION = "transaction",
}